# docs/openapi.json
{"openapi":"3.1.0","info":{"title":"AI-Driven Patrolling & Bandobast API","description":"\n## Overview\nREST API for police patrol management, event detection, and intelligent alerting.\n\n## Authentication\nAll endpoints require an API key passed via `X-API-Key` header.\n\n## Rate Limiting\n- 100 requests per minute per client\n- `X-RateLimit-Remaining` header shows remaining quota\n\n## Error Codes\n| Code | Description |\n|------|-------------|\n| 400 | Invalid request data |\n| 401 | Missing API key |\n| 403 | Invalid API key |\n| 404 | Resource not found |\n| 429 | Rate limit exceeded |\n| 500 | Internal server error |\n","version":"1.0.0"},"paths":{"/api/v1/events/ingest":{"post":{"tags":["Events"],"summary":"Ingest Event","operationId":"ingest_event_api_v1_events_ingest_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EventIngestRequest"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/EventResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/events":{"get":{"tags":["Events"],"summary":"List Events","operationId":"list_events_api_v1_events_get","parameters":[{"name":"camera_id","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Camera Id"}},{"name":"event_type","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Event Type"}},{"name":"start_date","in":"query","required":false,"schema":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Start Date"}},{"name":"end_date","in":"query","required":false,"schema":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"End Date"}},{"name":"processed","in":"query","required":false,"schema":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Processed"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","minimum":1,"default":1,"title":"Page"}},{"name":"per_page","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"default":20,"title":"Per Page"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PaginatedResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/events/{event_id}":{"get":{"tags":["Events"],"summary":"Get Event","operationId":"get_event_api_v1_events__event_id__get","parameters":[{"name":"event_id","in":"path","required":true,"schema":{"type":"integer","title":"Event Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/EventResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/events/{event_id}/process":{"patch":{"tags":["Events"],"summary":"Mark Processed","operationId":"mark_processed_api_v1_events__event_id__process_patch","parameters":[{"name":"event_id","in":"path","required":true,"schema":{"type":"integer","title":"Event Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/alerts":{"get":{"tags":["Alerts"],"summary":"List Alerts","operationId":"list_alerts_api_v1_alerts_get","parameters":[{"name":"severity","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Severity"}},{"name":"acknowledged","in":"query","required":false,"schema":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Acknowledged"}},{"name":"start_date","in":"query","required":false,"schema":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Start Date"}},{"name":"end_date","in":"query","required":false,"schema":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"End Date"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","minimum":1,"default":1,"title":"Page"}},{"name":"per_page","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"default":20,"title":"Per Page"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PaginatedResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/alerts/stats":{"get":{"tags":["Alerts"],"summary":"Alert Stats","operationId":"alert_stats_api_v1_alerts_stats_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/alerts/{alert_id}":{"get":{"tags":["Alerts"],"summary":"Get Alert","operationId":"get_alert_api_v1_alerts__alert_id__get","parameters":[{"name":"alert_id","in":"path","required":true,"schema":{"type":"integer","title":"Alert Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AlertResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["Alerts"],"summary":"Delete Alert","operationId":"delete_alert_api_v1_alerts__alert_id__delete","parameters":[{"name":"alert_id","in":"path","required":true,"schema":{"type":"integer","title":"Alert Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/alerts/{alert_id}/acknowledge":{"post":{"tags":["Alerts"],"summary":"Acknowledge Alert","operationId":"acknowledge_alert_api_v1_alerts__alert_id__acknowledge_post","parameters":[{"name":"alert_id","in":"path","required":true,"schema":{"type":"integer","title":"Alert Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AlertAcknowledgeRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AlertResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/patrol/start":{"post":{"tags":["Patrol"],"summary":"Start Patrol","operationId":"start_patrol_api_v1_patrol_start_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PatrolStartRequest"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PatrolSessionResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/patrol/end":{"post":{"tags":["Patrol"],"summary":"End Patrol","operationId":"end_patrol_api_v1_patrol_end_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PatrolEndRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PatrolSessionResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/patrol/sessions":{"get":{"tags":["Patrol"],"summary":"List Sessions","operationId":"list_sessions_api_v1_patrol_sessions_get","parameters":[{"name":"officer_id","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Officer Id"}},{"name":"status","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Status"}},{"name":"start_date","in":"query","required":false,"schema":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Start Date"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","minimum":1,"default":1,"title":"Page"}},{"name":"per_page","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"default":20,"title":"Per Page"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PaginatedResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/patrol/{session_id}":{"get":{"tags":["Patrol"],"summary":"Get Session","operationId":"get_session_api_v1_patrol__session_id__get","parameters":[{"name":"session_id","in":"path","required":true,"schema":{"type":"integer","title":"Session Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PatrolSessionResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/patrol/{session_id}/event":{"post":{"tags":["Patrol"],"summary":"Add Patrol Event","operationId":"add_patrol_event_api_v1_patrol__session_id__event_post","parameters":[{"name":"session_id","in":"path","required":true,"schema":{"type":"integer","title":"Session Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PatrolEventRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/summaries/generate":{"post":{"tags":["Summaries"],"summary":"Generate Summary","operationId":"generate_summary_api_v1_summaries_generate_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SummaryGenerateRequest"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SummaryResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/summaries":{"get":{"tags":["Summaries"],"summary":"List Summaries","operationId":"list_summaries_api_v1_summaries_get","parameters":[{"name":"summary_type","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Summary Type"}},{"name":"start_date","in":"query","required":false,"schema":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Start Date"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","minimum":1,"default":1,"title":"Page"}},{"name":"per_page","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"default":20,"title":"Per Page"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PaginatedResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/summaries/{summary_id}":{"get":{"tags":["Summaries"],"summary":"Get Summary","operationId":"get_summary_api_v1_summaries__summary_id__get","parameters":[{"name":"summary_id","in":"path","required":true,"schema":{"type":"integer","title":"Summary Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SummaryResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/cameras":{"post":{"tags":["Cameras"],"summary":"Create Camera","operationId":"create_camera_api_v1_cameras_post","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CameraCreateRequest"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CameraResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["Cameras"],"summary":"List Cameras","operationId":"list_cameras_api_v1_cameras_get","parameters":[{"name":"status","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Status"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","minimum":1,"default":1,"title":"Page"}},{"name":"per_page","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"default":20,"title":"Per Page"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PaginatedResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/cameras/{camera_id}":{"get":{"tags":["Cameras"],"summary":"Get Camera","operationId":"get_camera_api_v1_cameras__camera_id__get","parameters":[{"name":"camera_id","in":"path","required":true,"schema":{"type":"integer","title":"Camera Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CameraResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["Cameras"],"summary":"Delete Camera","operationId":"delete_camera_api_v1_cameras__camera_id__delete","parameters":[{"name":"camera_id","in":"path","required":true,"schema":{"type":"integer","title":"Camera Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/cameras/{camera_id}/status":{"patch":{"tags":["Cameras"],"summary":"Update Camera Status","operationId":"update_camera_status_api_v1_cameras__camera_id__status_patch","parameters":[{"name":"camera_id","in":"path","required":true,"schema":{"type":"integer","title":"Camera Id"}},{"name":"status","in":"query","required":true,"schema":{"type":"string","title":"Status"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/health":{"get":{"tags":["System"],"summary":"Health check","description":"Returns API health status","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1":{"get":{"tags":["System"],"summary":"API info","description":"Returns available endpoints","operationId":"api_info_api_v1_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"AlertAcknowledgeRequest":{"properties":{"acknowledged_by":{"type":"string","maxLength":100,"minLength":1,"title":"Acknowledged By"}},"type":"object","required":["acknowledged_by"],"title":"AlertAcknowledgeRequest"},"AlertResponse":{"properties":{"id":{"type":"integer","title":"Id"},"event_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Event Id"},"alert_type":{"type":"string","title":"Alert Type"},"severity":{"type":"string","title":"Severity"},"message":{"type":"string","title":"Message"},"location_lat":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Location Lat"},"location_lon":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Location Lon"},"acknowledged":{"type":"boolean","title":"Acknowledged"},"acknowledged_by":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Acknowledged By"},"acknowledged_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Acknowledged At"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["id","event_id","alert_type","severity","message","location_lat","location_lon","acknowledged","acknowledged_by","acknowledged_at","created_at"],"title":"AlertResponse"},"CameraCreateRequest":{"properties":{"camera_name":{"type":"string","maxLength":100,"minLength":1,"title":"Camera Name"},"location_name":{"type":"string","maxLength":200,"minLength":1,"title":"Location Name"},"latitude":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Latitude"},"longitude":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Longitude"}},"type":"object","required":["camera_name","location_name"],"title":"CameraCreateRequest"},"CameraResponse":{"properties":{"id":{"type":"integer","title":"Id"},"camera_name":{"type":"string","title":"Camera Name"},"location_name":{"type":"string","title":"Location Name"},"latitude":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Latitude"},"longitude":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Longitude"},"status":{"type":"string","title":"Status"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["id","camera_name","location_name","latitude","longitude","status","created_at"],"title":"CameraResponse"},"EventIngestRequest":{"properties":{"camera_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Camera Id"},"timestamp":{"type":"string","format":"date-time","title":"Timestamp"},"event_type":{"$ref":"#/components/schemas/EventTypeSchema"},"confidence_score":{"type":"number","maximum":1.0,"minimum":0.0,"title":"Confidence Score","default":0.5},"data":{"additionalProperties":true,"type":"object","title":"Data"}},"type":"object","required":["timestamp","event_type"],"title":"EventIngestRequest","example":{"camera_id":1,"confidence_score":0.85,"data":{"count":45,"density":2.3},"event_type":"crowd_detected","timestamp":"2026-01-16T21:00:00Z"}},"EventResponse":{"properties":{"id":{"type":"integer","title":"Id"},"camera_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Camera Id"},"timestamp":{"type":"string","format":"date-time","title":"Timestamp"},"event_type":{"type":"string","title":"Event Type"},"confidence_score":{"type":"number","title":"Confidence Score"},"data":{"additionalProperties":true,"type":"object","title":"Data"},"processed":{"type":"boolean","title":"Processed"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["id","camera_id","timestamp","event_type","confidence_score","data","processed","created_at"],"title":"EventResponse"},"EventTypeSchema":{"type":"string","enum":["object_detected","crowd_detected","static_object","crowd_surge","route_blocked"],"title":"EventTypeSchema"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"PaginatedResponse":{"properties":{"items":{"items":{},"type":"array","title":"Items"},"total":{"type":"integer","title":"Total"},"page":{"type":"integer","title":"Page"},"per_page":{"type":"integer","title":"Per Page"},"pages":{"type":"integer","title":"Pages"}},"type":"object","required":["items","total","page","per_page","pages"],"title":"PaginatedResponse"},"PatrolEndRequest":{"properties":{"session_id":{"type":"integer","title":"Session Id"},"final_location":{"anyOf":[{"additionalProperties":{"type":"number"},"type":"object"},{"type":"null"}],"title":"Final Location"}},"type":"object","required":["session_id"],"title":"PatrolEndRequest"},"PatrolEventRequest":{"properties":{"event_type":{"type":"string","title":"Event Type"},"description":{"type":"string","title":"Description"},"location":{"anyOf":[{"additionalProperties":{"type":"number"},"type":"object"},{"type":"null"}],"title":"Location"}},"type":"object","required":["event_type","description"],"title":"PatrolEventRequest"},"PatrolSessionResponse":{"properties":{"id":{"type":"integer","title":"Id"},"officer_id":{"type":"string","title":"Officer Id"},"officer_name":{"type":"string","title":"Officer Name"},"start_time":{"type":"string","format":"date-time","title":"Start Time"},"end_time":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"End Time"},"status":{"type":"string","title":"Status"},"incidents_count":{"type":"integer","title":"Incidents Count"},"distance_km":{"type":"number","title":"Distance Km"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["id","officer_id","officer_name","start_time","end_time","status","incidents_count","distance_km","created_at"],"title":"PatrolSessionResponse"},"PatrolStartRequest":{"properties":{"officer_id":{"type":"string","maxLength":50,"minLength":1,"title":"Officer Id"},"officer_name":{"type":"string","maxLength":100,"minLength":1,"title":"Officer Name"},"initial_location":{"anyOf":[{"additionalProperties":{"type":"number"},"type":"object"},{"type":"null"}],"title":"Initial Location"}},"type":"object","required":["officer_id","officer_name"],"title":"PatrolStartRequest","example":{"initial_location":{"lat":28.6139,"lon":77.209},"officer_id":"OFF001","officer_name":"Inspector Singh"}},"SummaryGenerateRequest":{"properties":{"summary_type":{"$ref":"#/components/schemas/SummaryTypeSchema"},"reference_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Reference Id"},"reference_date":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Reference Date"}},"type":"object","required":["summary_type"],"title":"SummaryGenerateRequest"},"SummaryResponse":{"properties":{"id":{"type":"integer","title":"Id"},"summary_type":{"type":"string","title":"Summary Type"},"patrol_session_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Patrol Session Id"},"reference_date":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Reference Date"},"content":{"type":"string","title":"Content"},"key_insights":{"items":{},"type":"array","title":"Key Insights"},"risk_score":{"type":"number","title":"Risk Score"},"generated_at":{"type":"string","format":"date-time","title":"Generated At"}},"type":"object","required":["id","summary_type","patrol_session_id","reference_date","content","key_insights","risk_score","generated_at"],"title":"SummaryResponse"},"SummaryTypeSchema":{"type":"string","enum":["patrol","bandobast","daily"],"title":"SummaryTypeSchema"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}